gcc3: Fix bug in collect_execute(): 'open' with O_CREAT needs third argument.
The collect_execute() function in gcc/collect2.c calls 'open' with
the 'O_CREAT' flag, but the third argument is missing.  This is
equivalent to the problem described in Gentoo bug #256638
(see <http://bugs.gentoo.org/256638>) and the patch there applies
here.

Note: This is for the nios2 no-mmu toolchain.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>